refactor(arch): close tach blind spot — declare loop/docker, break core↔loop cycle
The review found tach.toml silently omitted teatree.loop (1.4k LOC,
cross-layer, heavily churned) and teatree.docker, so the dependency
guardrail could not constrain the subsystem absorbing the most change.

- Declare teatree.loop + teatree.docker as tach modules with real deps.
- Model teatree.core.management as a composition-root/entrypoint layer
  (like teatree.cli): mgmt commands legitimately wire loop/docker, but
  teatree.core (domain) must not — that is the real layering invariant.
- Invert the two genuine domain smells the blind spot hid:
  * _find_overlay_db: loop.scanners -> teatree.paths.find_overlay_db.
  * image_tag_for_lockfile(cfg) -> BaseImageConfig.image_tag() method.
- cli depends_on += loop (cli is the top entrypoint layer).

tach now validates a complete, acyclic core graph. 577 tests green.
