+++ src/teatree/docker/__init__.py
@@ -1,5 +1,5 @@
 """Docker helpers — base-image sharing, compose orchestration."""
 
-from teatree.docker.build import ensure_base_image, image_tag_for_lockfile
+from teatree.docker.build import ensure_base_image
 
-__all__ = ["ensure_base_image", "image_tag_for_lockfile"]
+__all__ = ["ensure_base_image"]
